API work has begun
Basic API structure has been added to app.py (it is currently block commented between lines 45 and 59). The aim of this API has been decided to supply the user with probabilities from various probability distributions so work has begun on a Poisson distribution function, utilising a decorator to ensure that the correct number of arguments are supplied.