Continued building of Poisson function
Logging has now been implemented and the Poisson function now checks that the correct data types are used. This needs to be improved with increased specificity - currently it allows all floats and ints but this isn't correct. Further work needs to be done to add other calculations to the function (so far the only working function is that for probability).