Catching overflow errors
If the numbers involved in the calculation are too big, overflow errors are produced which halts the calculation and returns an "internal server error" from the API even if the request is reasonable. Try and except statements have been put into all Poisson functions to catch this and return an error message from the API while still returning a 200 response.